Began working on interactions

+++ scripts/item/Weapon.cs
@@ -2,6 +2,6 @@ namespace LandsOfAzerith.scripts.item;
 
 public abstract class Weapon : Item
 {
-    public abstract ulong Damage { get; }
-    public abstract ulong Range { get; }
+    public abstract uint Damage { get; }
+    public abstract uint Range { get; }
 }
